Refactor skelenox in a package
